Refactor hook test code
