+++ tests/dx12.rs
@@ -1,40 +1,17 @@
 mod harness;
+mod hook;
 
 use std::thread;
 use std::time::Duration;
 
 use harness::dx12::Dx12Harness;
+use hook::HookExample;
 use hudhook::hooks::dx12::ImguiDx12Hooks;
-use hudhook::hooks::{self, ImguiRenderLoop};
-use imgui::Condition;
+use hudhook::hooks::{Hooks, ImguiRenderLoop};
 use tracing::metadata::LevelFilter;
 
 #[test]
 fn test_imgui_dx12() {
-    struct Dx12HookExample;
-
-    impl Dx12HookExample {
-        fn new() -> Self {
-            println!("Initializing");
-            hudhook::utils::alloc_console();
-
-            Dx12HookExample
-        }
-    }
-
-    impl ImguiRenderLoop for Dx12HookExample {
-        fn render(&mut self, ui: &mut imgui::Ui) {
-            ui.window("Hello world").size([300.0, 300.0], Condition::FirstUseEver).build(|| {
-                ui.text("Hello world!");
-                ui.text("こんにちは世界！");
-                ui.text("This...is...imgui-rs!");
-                ui.separator();
-                let mouse_pos = ui.io().mouse_pos;
-                ui.text(format!("Mouse Position: ({:.1},{:.1})", mouse_pos[0], mouse_pos[1]));
-            });
-        }
-    }
-
     tracing_subscriber::fmt()
         .with_max_level(LevelFilter::TRACE)
         .with_thread_ids(true)
@@ -47,7 +24,7 @@ fn test_imgui_dx12() {
     thread::sleep(Duration::from_millis(500));
 
     unsafe {
-        let hooks: Box<dyn hooks::Hooks> = { Dx12HookExample::new().into_hook::<ImguiDx12Hooks>() };
+        let hooks: Box<dyn Hooks> = { HookExample::new().into_hook::<ImguiDx12Hooks>() };
         hooks.hook();
         hudhook::lifecycle::global_state::set_hooks(hooks);
     }
